requirements.txt
# Core dependencies
python-dotenv==1.0.0
pandas==2.1.4
numpy>=1.23.2,<1.26
scikit-learn==1.3.2
xgboost==2.0.2

# Data processing and analysis
dbt-core==1.7.4
dbt-postgres==1.7.4
ydata-profiling==4.6.4
sweetviz==2.2.1
faker==20.1.0

# Dashboard and visualization
streamlit==1.29.0
plotly==5.18.0
seaborn==0.12.2
matplotlib==3.8.2

# Database
duckdb==0.9.2
psycopg2-binary==2.9.9

# Testing and development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
jupyter==1.0.0

# ML and model evaluation
shap==0.43.0
optuna==3.4.0
# joblib version managed by scikit-learn

# Utilities
tqdm==4.66.1
requests==2.31.0
python-dateutil==2.8.2 